Remove substitution_formatter_base.h
