+++ source/common/matcher/actions/string_returning_action.cc
@@ -2,7 +2,7 @@
 
 #include "envoy/config/core/v3/substitution_format_string.pb.h"
 #include "envoy/config/core/v3/substitution_format_string.pb.validate.h"
-#include "envoy/formatter/substitution_formatter_base.h"
+#include "envoy/formatter/substitution_formatter.h"
 #include "envoy/registry/registry.h"
 
 #include "source/common/formatter/substitution_format_string.h"
