+++ source/common/router/header_parser.h
@@ -5,7 +5,7 @@
 
 #include "envoy/access_log/access_log.h"
 #include "envoy/config/core/v3/base.pb.h"
-#include "envoy/formatter/substitution_formatter_base.h"
+#include "envoy/formatter/substitution_formatter.h"
 #include "envoy/http/header_evaluator.h"
 #include "envoy/http/header_map.h"
